implement model for cardsets seasons endpoint
adjusted automated tests as well
